requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
Django==4.2.5
django-bootstrap4==23.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
methodism==0.2.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.0
tzdata==2023.3
urllib3==2.0.4
whitenoise==6.5.0
